+++ Braintree/TransactionRequest.cs
@@ -49,8 +49,9 @@ namespace Braintree
         public String PaymentMethodToken { get; set; }
         public String CustomerId { get; set; }
         public String ShippingAddressId { get; set; }
+        public String BillingAddressId { get; set; }
         public String VenmoSdkPaymentMethodCode { get; set; }
-        public Decimal ServiceFeeAmount { get; set; }
+        public Decimal? ServiceFeeAmount { get; set; }
 
         public TransactionRequest()
         {
@@ -97,12 +98,13 @@ namespace Braintree
             builder.AddElement("payment-method-token", PaymentMethodToken);
             builder.AddElement("purchase-order-number", PurchaseOrderNumber);
             builder.AddElement("shipping-address-id", ShippingAddressId);
+            builder.AddElement("billing-address-id", BillingAddressId);
             if (TaxAmount != 0) builder.AddElement("tax-amount", TaxAmount);
             if (TaxExempt.HasValue) {
                 builder.AddElement("tax-exempt", TaxExempt);
             }
             builder.AddElement("merchant-account-id", MerchantAccountId);
-            if (ServiceFeeAmount != 0) builder.AddElement("service-fee-amount", ServiceFeeAmount);
+            if (ServiceFeeAmount.HasValue) builder.AddElement("service-fee-amount", ServiceFeeAmount);
 
             if (Type != null) builder.AddElement("type", Type.ToString().ToLower());